Various fine tuning fixes.
The nav bar when the window is scrolled decrease size.
Some corrections to the demo section buttons.
Change the style of the demo section to make it always visible.
Implement a carousel for the "itWork" section.
Implement post-image loading for "demo" section and "itWork" section.
Change tablet trigger from 791px to 990px.
Insert the "login" and the translate button in the scroll down mobile header.